refactor: 폴더 구조 변경 및 TODO 작성
