+++ src/main/java/com/wellmeet/favorite/service/FavoriteService.java
@@ -5,9 +5,9 @@ import com.wellmeet.member.domain.Member;
 import com.wellmeet.member.domain.MemberRestaurant;
 import com.wellmeet.member.service.MemberRestaurantService;
 import com.wellmeet.member.service.MemberService;
-import com.wellmeet.recommend.restaurant.domain.Restaurant;
-import com.wellmeet.recommend.restaurant.service.RestaurantService;
-import com.wellmeet.recommend.review.service.ReviewService;
+import com.wellmeet.restaurant.RestaurantService;
+import com.wellmeet.restaurant.domain.Restaurant;
+import com.wellmeet.restaurant.model.review.service.ReviewService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
