+++ src/main/java/com/wellmeet/member/repository/MemberRestaurantRepository.java
@@ -2,7 +2,7 @@ package com.wellmeet.member.repository;
 
 import com.wellmeet.member.domain.Member;
 import com.wellmeet.member.domain.MemberRestaurant;
-import com.wellmeet.recommend.restaurant.domain.Restaurant;
+import com.wellmeet.restaurant.domain.Restaurant;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
