+++ src/main/java/com/wellmeet/recommend/RecommendController.java
@@ -1,12 +1,10 @@
 package com.wellmeet.recommend;
 
 import com.wellmeet.recommend.crawlingreview.domain.VibeName;
-import com.wellmeet.recommend.restaurant.dto.RecommendRestaurantResponse;
-import com.wellmeet.recommend.restaurant.dto.RestaurantResponse;
+import com.wellmeet.recommend.dto.RecommendRestaurantResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,21 +21,5 @@ public class RecommendController {
             @RequestParam(value = "longitude") double longitude
     ) {
         return recommendService.getRecommendRestaurants(vibeName, latitude, longitude);
-    }
-
-    @GetMapping("/api/restaurants/nearby")
-    public List<RecommendRestaurantResponse> getNearbyRestaurants(
-            @RequestParam(value = "latitude") double latitude,
-            @RequestParam(value = "longitude") double longitude
-    ) {
-        return recommendService.getNearbyRestaurants(latitude, longitude);
-    }
-
-    @GetMapping("/api/restaurant/{id}")
-    public RestaurantResponse getRestaurant(
-            @RequestParam(value = "memberId") Long memberId,
-            @PathVariable Long id
-    ) {
-        return recommendService.getRestaurant(id, memberId);
     }
 }
