--- src/main/java/com/wellmeet/recommend/menu/repository/MenuRepository.java
@@ -1,12 +1,0 @@
-package com.wellmeet.recommend.menu.repository;
-
-import com.wellmeet.recommend.menu.domain.Menu;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MenuRepository extends JpaRepository<Menu, Long> {
-
-    List<Menu> findByRestaurantId(Long restaurantId);
-}
