--- src/main/java/com/wellmeet/recommend/restaurant/dto/RepresentativeReviewResponse.java
@@ -1,20 +1,0 @@
-package com.wellmeet.recommend.restaurant.dto;
-
-import com.wellmeet.recommend.review.domain.Review;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class RepresentativeReviewResponse {
-
-    private String situation;
-    private String content;
-    private String logo;
-
-    public RepresentativeReviewResponse(Review review) {
-        this.situation = review.getSituation().getName();
-        this.content = review.getContent();
-        this.logo = review.getSituation().getLogo();
-    }
-}
