--- src/main/java/com/wellmeet/recommend/review/repository/ReviewRepository.java
@@ -1,12 +1,0 @@
-package com.wellmeet.recommend.review.repository;
-
-import com.wellmeet.recommend.review.domain.Review;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-
-    List<Review> findByRestaurantId(Long restaurantId);
-}
