+++ src/main/java/com/wellmeet/restaurant/domain/Restaurant.java
@@ -1,0 +1,47 @@
+package com.wellmeet.restaurant.domain;
+
+import com.wellmeet.common.domain.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Restaurant extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @NotNull
+    @Column(unique = true)
+    private String placeId;
+
+    @NotBlank
+    private String name;
+
+    @NotBlank
+    private String address;
+
+    private double latitude;
+    private double longitude;
+    private String thumbnail;
+
+    public Restaurant(String name, String address, double latitude, double longitude, String thumbnail,
+                      String placeId) {
+        this.name = name;
+        this.address = address;
+        this.latitude = latitude;
+        this.longitude = longitude;
+        this.thumbnail = thumbnail;
+        this.placeId = placeId;
+    }
+}
