+++ src/main/java/com/wellmeet/restaurant/dto/NearbyRestaurantResponse.java
@@ -1,4 +1,4 @@
-package com.wellmeet.favorite.dto;
+package com.wellmeet.restaurant.dto;
 
 import com.wellmeet.restaurant.domain.Restaurant;
 import lombok.Getter;
@@ -6,18 +6,20 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class FavoriteRestaurantResponse {
+public class NearbyRestaurantResponse {
 
     private Long id;
     private String name;
     private String address;
+    private double distance;
     private double rating;
     private String thumbnail;
 
-    public FavoriteRestaurantResponse(Restaurant restaurant, double rating) {
+    public NearbyRestaurantResponse(Restaurant restaurant, double distance, double rating) {
         this.id = restaurant.getId();
         this.name = restaurant.getName();
         this.address = restaurant.getAddress();
+        this.distance = distance;
         this.rating = rating;
         this.thumbnail = restaurant.getThumbnail();
     }
