+++ src/main/java/com/wellmeet/restaurant/model/menu/domain/Menu.java
@@ -1,5 +1,6 @@
-package com.wellmeet.recommend.crawlingreview.domain;
+package com.wellmeet.restaurant.model.menu.domain;
 
+import com.wellmeet.common.domain.BaseEntity;
 import com.wellmeet.restaurant.domain.Restaurant;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,8 +9,8 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,26 +18,27 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CrawlingReview { // TODO : MSA를 도입하면서 추천 서버를 분리 후 recommend 패키지 삭제
+public class Menu extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @NotNull
-    private String content;
-
-    @NotNull
-    private LocalDateTime createdAt;
+    @NotBlank
+    private String name;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "restaurant_id")
     private Restaurant restaurant;
 
-    public CrawlingReview(String content, Restaurant restaurant) {
-        this.content = content;
+    private String description;
+    private int price;
+
+    public Menu(String name, String description, int price, Restaurant restaurant) {
+        this.name = name;
+        this.description = description;
+        this.price = price;
         this.restaurant = restaurant;
-        this.createdAt = LocalDateTime.now();
     }
 }
