+++ src/main/java/com/wellmeet/restaurant/model/menu/repository/MenuRepository.java
@@ -1,0 +1,12 @@
+package com.wellmeet.restaurant.model.menu.repository;
+
+import com.wellmeet.restaurant.model.menu.domain.Menu;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+    List<Menu> findByRestaurantId(Long restaurantId);
+}
