+++ src/main/java/com/wellmeet/restaurant/model/review/domain/ReviewTag.java
@@ -1,6 +1,5 @@
-package com.wellmeet.recommend.crawlingreview.domain;
+package com.wellmeet.restaurant.model.review.domain;
 
-import com.wellmeet.restaurant.domain.Restaurant;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -9,7 +8,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,26 +15,19 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CrawlingReview { // TODO : MSA를 도입하면서 추천 서버를 분리 후 recommend 패키지 삭제
+public class ReviewTag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotNull
-    private String content;
-
-    @NotNull
-    private LocalDateTime createdAt;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "review_id")
+    private Review review;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "restaurant_id")
-    private Restaurant restaurant;
-
-    public CrawlingReview(String content, Restaurant restaurant) {
-        this.content = content;
-        this.restaurant = restaurant;
-        this.createdAt = LocalDateTime.now();
-    }
+    @JoinColumn(name = "tag_id")
+    private Tag tag;
 }
