+++ src/main/java/com/wellmeet/restaurant/model/review/domain/Situation.java
@@ -1,0 +1,20 @@
+package com.wellmeet.restaurant.model.review.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum Situation {
+
+    DATE("데이트", "💕"),
+    FAMILY("가족 모임", "🎎"),
+    BUSINESS("비즈니스 미팅", "🧳"),
+    ;
+
+    private final String name;
+    private final String logo;
+
+    Situation(String name, String logo) {
+        this.name = name;
+        this.logo = logo;
+    }
+}
