+++ src/test/java/com/wellmeet/BaseControllerTest.java
@@ -3,10 +3,10 @@ package com.wellmeet;
 import com.wellmeet.member.repository.MemberRepository;
 import com.wellmeet.member.repository.MemberRestaurantRepository;
 import com.wellmeet.recommend.crawlingreview.repository.VibeRepository;
-import com.wellmeet.recommend.menu.repository.MenuRepository;
-import com.wellmeet.recommend.restaurant.repository.RestaurantRepository;
-import com.wellmeet.recommend.restaurant.tool.CrawlingReviewGenerator;
-import com.wellmeet.recommend.review.repository.ReviewRepository;
+import com.wellmeet.restaurant.model.menu.repository.MenuRepository;
+import com.wellmeet.restaurant.model.review.repository.ReviewRepository;
+import com.wellmeet.restaurant.repository.RestaurantRepository;
+import com.wellmeet.restaurant.tool.CrawlingReviewGenerator;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.filter.log.RequestLoggingFilter;
