+++ src/test/java/com/wellmeet/BaseRepositoryTest.java
@@ -2,7 +2,7 @@ package com.wellmeet;
 
 
 import com.wellmeet.config.JpaAuditingConfig;
-import com.wellmeet.recommend.restaurant.tool.CrawlingReviewGenerator;
+import com.wellmeet.restaurant.tool.CrawlingReviewGenerator;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
