+++ src/test/java/com/wellmeet/restaurant/controller/RestaurantControllerTest.java
@@ -1,0 +1,71 @@
+package com.wellmeet.restaurant.controller;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.wellmeet.BaseControllerTest;
+import com.wellmeet.member.domain.Member;
+import com.wellmeet.recommend.dto.RecommendRestaurantResponse;
+import com.wellmeet.restaurant.domain.Restaurant;
+import com.wellmeet.restaurant.dto.RestaurantResponse;
+import com.wellmeet.restaurant.model.menu.domain.Menu;
+import com.wellmeet.restaurant.model.review.domain.Review;
+import com.wellmeet.restaurant.model.review.domain.Situation;
+import io.restassured.http.ContentType;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
+
+class RestaurantControllerTest extends BaseControllerTest {
+
+    private static final double LATITUDE = 38.5;
+    private static final double LONGITUDE = 128.2;
+    private static final String THUMBNAIL = "https://example.com/restaurant.jpg";
+
+    @Test
+    @DisplayName("주변 레스토랑 조회")
+    void getNearbyRestaurants() {
+        Restaurant restaurant1 = new Restaurant("restaurant1", "address1", LATITUDE, LONGITUDE, THUMBNAIL, "123");
+        restaurantRepository.save(restaurant1);
+        Restaurant restaurant2 = new Restaurant("restaurant2", "address2", LATITUDE, LONGITUDE, THUMBNAIL, "124");
+        restaurantRepository.save(restaurant2);
+        Restaurant restaurant3 = new Restaurant("restaurant3", "address3", LATITUDE - 3, LONGITUDE + 3, THUMBNAIL,
+                "125");
+        restaurantRepository.save(restaurant3);
+
+        RecommendRestaurantResponse[] responses = given()
+                .contentType(ContentType.JSON)
+                .when().get("/api/restaurants/nearby?latitude=" + LATITUDE + "&longitude=" + LONGITUDE)
+                .then().statusCode(HttpStatus.OK.value())
+                .extract().as(RecommendRestaurantResponse[].class);
+
+        assertThat(responses).hasSize(2);
+    }
+
+    @Test
+    @DisplayName("레스토랑 상세 조회")
+    void getRestaurant() {
+        Restaurant restaurant = new Restaurant("restaurant1", "address1", LATITUDE, LONGITUDE, THUMBNAIL, "123");
+        restaurantRepository.save(restaurant);
+        Menu menu1 = new Menu("menu1", "description1", 10000, restaurant);
+        menuRepository.save(menu1);
+        Menu menu2 = new Menu("menu2", "description2", 15000, restaurant);
+        menuRepository.save(menu2);
+        Member member = new Member("nickname");
+        memberRepository.save(member);
+        Review review1 = new Review("review1", 5, Situation.DATE, restaurant, member);
+        reviewRepository.save(review1);
+        Review review2 = new Review("review2", 4, Situation.BUSINESS, restaurant, member);
+        reviewRepository.save(review2);
+
+        RestaurantResponse restaurantResponse = given()
+                .contentType(ContentType.JSON)
+                .queryParam("memberId", member.getId())
+                .when().get("/api/restaurant/{id}", restaurant.getId())
+                .then().statusCode(HttpStatus.OK.value())
+                .extract().as(RestaurantResponse.class);
+
+        assertThat(restaurantResponse.getId()).isEqualTo(restaurant.getId());
+        assertThat(restaurantResponse.getMenus()).hasSize(2);
+        assertThat(restaurantResponse.getReviews()).hasSize(2);
+    }
+}
