+++ src/test/java/com/wellmeet/restaurant/repository/RestaurantRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.wellmeet.recommend;
+package com.wellmeet.restaurant.repository;
 
 import static com.wellmeet.recommend.crawlingreview.domain.VibeName.CLASSIC;
 import static com.wellmeet.recommend.crawlingreview.domain.VibeName.CLEAN;
@@ -6,23 +6,30 @@ import static com.wellmeet.recommend.crawlingreview.domain.VibeName.LIVELY;
 import static com.wellmeet.recommend.crawlingreview.domain.VibeName.MODERN;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.wellmeet.BaseControllerTest;
+import com.wellmeet.BaseRepositoryTest;
 import com.wellmeet.recommend.crawlingreview.domain.Vibe;
 import com.wellmeet.recommend.crawlingreview.domain.VibeName;
-import com.wellmeet.recommend.dto.RecommendRestaurantResponse;
+import com.wellmeet.recommend.crawlingreview.repository.VibeRepository;
+import com.wellmeet.restaurant.domain.BoundingBox;
 import com.wellmeet.restaurant.domain.Restaurant;
-import io.restassured.http.ContentType;
 import java.util.Arrays;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
+import org.springframework.beans.factory.annotation.Autowired;
 
-class RecommendControllerTest extends BaseControllerTest {
+class RestaurantRepositoryTest extends BaseRepositoryTest {
 
     private static final double LATITUDE = 38.5;
     private static final double LONGITUDE = 128.2;
     private static final String THUMBNAIL = "https://example.com/restaurant.jpg";
+
+    @Autowired
+    private RestaurantRepository restaurantRepository;
+
+    @Autowired
+    protected VibeRepository vibeRepository;
 
     @BeforeEach
     void setEnvironment() {
@@ -32,7 +39,7 @@ class RecommendControllerTest extends BaseControllerTest {
 
     @Test
     @DisplayName("추천 레스토랑 조회 - vibe에 따른 비율로 정렬")
-    void getRecommendRestaurants() {
+    void findRestaurantsOrderedByVibeRatioWithBoundBox() {
         Restaurant restaurant1 = new Restaurant("restaurant1", "address1", LATITUDE, LONGITUDE, THUMBNAIL, "123");
         Restaurant savedRestaurant1 = restaurantRepository.save(restaurant1);
         createCrawlingReviews(savedRestaurant1, CLASSIC, CLASSIC, CLASSIC, CLEAN, LIVELY);
@@ -49,19 +56,32 @@ class RecommendControllerTest extends BaseControllerTest {
         Restaurant savedRestaurant4 = restaurantRepository.save(restaurant4);
         createCrawlingReviews(savedRestaurant4, LIVELY, LIVELY, LIVELY);
 
-        RecommendRestaurantResponse[] responses = given()
-                .contentType(ContentType.JSON)
-                .when().get("/api/restaurants/recommend?" +
-                        "vibe=" + CLASSIC.name() +
-                        "&latitude=" + LATITUDE +
-                        "&longitude=" + LONGITUDE)
-                .then().statusCode(HttpStatus.OK.value())
-                .extract().as(RecommendRestaurantResponse[].class);
+        BoundingBox boundingBox = new BoundingBox(LATITUDE, LONGITUDE);
+        List<Restaurant> restaurantsOrderedByVibeRatio = restaurantRepository.findRestaurantsOrderedByVibeRatioWithBoundBox(
+                CLASSIC, boundingBox
+        );
 
-        assertThat(responses).hasSize(3);
-        assertThat(responses[0].getId()).isEqualTo(savedRestaurant3.getId());
-        assertThat(responses[1].getId()).isEqualTo(savedRestaurant1.getId());
-        assertThat(responses[2].getId()).isEqualTo(savedRestaurant2.getId());
+        assertThat(restaurantsOrderedByVibeRatio).hasSize(3);
+        assertThat(restaurantsOrderedByVibeRatio.getFirst().getId()).isEqualTo(savedRestaurant3.getId());
+        assertThat(restaurantsOrderedByVibeRatio.get(1).getId()).isEqualTo(savedRestaurant1.getId());
+        assertThat(restaurantsOrderedByVibeRatio.get(2).getId()).isEqualTo(savedRestaurant2.getId());
+    }
+
+    @Test
+    @DisplayName("주변 레스토랑 조회 - BoundingBox를 이용한 레스토랑 조회")
+    void findWithBoundBox() {
+        Restaurant restaurant1 = new Restaurant("restaurant1", "address1", LATITUDE, LONGITUDE, THUMBNAIL, "123");
+        restaurantRepository.save(restaurant1);
+        Restaurant restaurant2 = new Restaurant("restaurant2", "address2", LATITUDE, LONGITUDE, THUMBNAIL, "124");
+        restaurantRepository.save(restaurant2);
+        Restaurant restaurant3 = new Restaurant("restaurant3", "address3", LATITUDE - 3, LONGITUDE + 3, THUMBNAIL,
+                "125");
+        restaurantRepository.save(restaurant3);
+
+        BoundingBox boundingBox = new BoundingBox(LATITUDE, LONGITUDE);
+        List<Restaurant> restaurants = restaurantRepository.findWithBoundBox(boundingBox);
+
+        assertThat(restaurants).hasSize(2);
     }
 
     private void createCrawlingReviews(Restaurant restaurant, VibeName... vibeNames) {
